[HOTSPOT-210] refactor: 조회 전용 dto 디렉토리 이동
