--- src/main/java/hotspot/admin/family/domain/FamilyListSortType.java
@@ -1,8 +1,0 @@
-package hotspot.admin.family.domain;
-
-public enum FamilyListSortType {
-    FAMILY_ID,
-    MEMBER_COUNT,
-    NAME,
-    USAGE
-}
