+++ src/main/java/hotspot/admin/family/infrastructure/query/dto/FamilyApprovalTargetInfo.java
@@ -1,10 +1,10 @@
-package hotspot.admin.family.service.dto;
+package hotspot.admin.family.infrastructure.query.dto;
 
 import hotspot.admin.family.domain.FamilyRole;
 import lombok.Builder;
 
 @Builder
-public record FamilyAddApprovalInfo(
+public record FamilyApprovalTargetInfo(
         Long familyId,
         Long targetSubId,
         FamilyRole targetFamilyRole
