+++ src/main/java/hotspot/admin/family/infrastructure/query/dto/FamilyPolicyMemberRow.java
@@ -1,4 +1,4 @@
-package hotspot.admin.family.service.dto;
+package hotspot.admin.family.infrastructure.query.dto;
 
 import hotspot.admin.family.domain.FamilyRole;
 import lombok.Builder;
