+++ src/main/java/hotspot/admin/family/infrastructure/query/dto/FamilyPolicyStatusRow.java
@@ -1,4 +1,4 @@
-package hotspot.admin.family.service.dto;
+package hotspot.admin.family.infrastructure.query.dto;
 
 import java.util.List;
