+++ src/main/java/hotspot/admin/family/service/GetFamilyPolicyStatusServiceImpl.java
@@ -12,7 +12,7 @@ import hotspot.admin.common.util.PhoneCryptoUtil;
 import hotspot.admin.common.util.PhoneMaskingUtil;
 import hotspot.admin.family.controller.port.GetFamilyPolicyStatusService;
 import hotspot.admin.family.controller.response.FamilyPolicyMemberStatusItem;
-import hotspot.admin.family.service.dto.FamilyPolicyStatusRow;
+import hotspot.admin.family.infrastructure.query.dto.FamilyPolicyStatusRow;
 import hotspot.admin.family.service.port.FamilyRepository;
 import hotspot.admin.family.service.port.FamilySubQueryRepository;
 import lombok.RequiredArgsConstructor;
