+++ src/main/java/hotspot/admin/family/service/port/FamilySubQueryRepository.java
@@ -2,11 +2,11 @@ package hotspot.admin.family.service.port;
 
 import java.util.List;
 
-import hotspot.admin.family.service.dto.FamilyControlMemberRow;
-import hotspot.admin.family.service.dto.FamilyPolicyAppPolicyRow;
-import hotspot.admin.family.service.dto.FamilyPolicyMemberRow;
-import hotspot.admin.family.service.dto.FamilyPolicyStatusRow;
-import hotspot.admin.family.service.dto.FamilyPolicyTimePolicyRow;
+import hotspot.admin.family.infrastructure.query.dto.FamilyControlMemberRow;
+import hotspot.admin.family.infrastructure.query.dto.FamilyPolicyAppPolicyRow;
+import hotspot.admin.family.infrastructure.query.dto.FamilyPolicyMemberRow;
+import hotspot.admin.family.infrastructure.query.dto.FamilyPolicyStatusRow;
+import hotspot.admin.family.infrastructure.query.dto.FamilyPolicyTimePolicyRow;
 
 public interface FamilySubQueryRepository {
     List<FamilyControlMemberRow> findFamilyControlMembers(Long familyId);
